Fix gui_window example compilation
I remember making this fix but it got lost somewhere between my plethora
of git branches and stashes. Like tears in rain.

Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/135915
Tested-by: Continuous Integration <build@kdab.com>

+++ examples/gui_window/gui_window.cpp
@@ -18,7 +18,7 @@ class ExampleWindow : public KDGui::Window
     {
         spdlog::info("{}() buttons = {} at pos = ({}, {})",
                      __FUNCTION__,
-                     ev->buttons(),
+                     ev->buttons().toInt(),
                      ev->xPos(),
                      ev->yPos());
     }
@@ -27,7 +27,7 @@ class ExampleWindow : public KDGui::Window
     {
         spdlog::info("{}() buttons = {} at pos = ({}, {})",
                      __FUNCTION__,
-                     ev->buttons(),
+                     ev->buttons().toInt(),
                      ev->xPos(),
                      ev->yPos());
     }
@@ -36,7 +36,7 @@ class ExampleWindow : public KDGui::Window
     {
         spdlog::info("{}() buttons = {} at pos = ({}, {})",
                      __FUNCTION__,
-                     ev->buttons(),
+                     ev->buttons().toInt(),
                      ev->xPos(),
                      ev->yPos());
     }
